Migrate check_link_target_fallback to BookTest
